+++ PowerShell/Invoke-CCMSetupBulkRegistrationToken/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Invoke-CCMSetupBulkRegistrationToken release history
 
+## 1.3.0 - 2022-12-05
+
+* Added `Workroup Joined` parameters.
+* Re-wrote documentation in markdown.
+* Added server side task parameters to re-run if failed.
+* Changed `NT SYSTEM` to SID in order for fix script for non english locales
+
 ## 1.2.0 - 2022-08-04
 
 * Fixed Parameter sets in `New-CMClientBulkRegistrationToken`